requirements.txt
rich
beautifulsoup4
pycrypto
requests
python-qbittorrent
python-dotenv